PB-1009: trust API Gateway to authenticate users.
We are delegating the authentication to API Gateway which sets the
`Geoadmin-Username` header. However due to how API Gateway and JWT-based
authentication work, the header is only set at login time. It is on the service
to keep track of the user afterwards.

This change updates service-stac in Dev to trust the `Geoadmin-Username` header
if it is present. Then service-stac persists the user across their whole session.

Relevant documentation is at https://docs.djangoproject.com/en/5.1/howto/auth-remote-user/