Set default file path for assets
When assets are created via admin GUI there is no file path set. If the file is
then uploaded via API using the multipart upload, the file path will never be
set on the asset.

This aligns the admin GUI behaviour with that of creating an asset via API PUT.