PB-1073 Make properties visible again in OpenAPI docs
Before this, the "properties" field was not visible in the request body schema of `Data-Management/operation/putFeature`.

This happened by accident when the whole "properties" field was set to `readOnly: true` in commit 613f084 (context: PB-687 to update the openapi validator).

You can see the bug also in production when you compare v0.9 (which apparently is not updated anymore) and the newer v1.0. The non-forecast fields visible now (`datetime`, `start_datetime`, `end_datetime`, `title`) are taken from the v0.9 version. So `created` and `expired` are made invisible.

Implementation notes:
- Properties `created` and `updated` are set to `readOnly` in the referenced definition already.
- OpenAPI <3.1 ^does not support sibling keywords alongside $ref, the workaround is to use `allOf`.