PB-1238: include static files in debug image.
The debug Docker image does not include the static files. This means that when
we run it in Kubernetes (in the dev staging), it fails to serve some requests.
This make debugging and testing harder than it needs to be.

This change updates Dockerfile to include the static files in the debug image.