PB-1073 Mark "expires" as writeable in the API docs
In PB-687, this had accidentally been set to read-only, so it disappeared in the request body of putFeature. In the context of PB-1073, we set this again to read-only because it did not show in the released v0.9 docs. This is because "expires" was only added later in PB-847. So we set it to writable again, as was the original intent.