PB-1252: store file sizes of assets and collection total data size as BigIntegers
Currently there's an issue on PROD when uploading large files. In this case it was 12.5 GB.
The file sizes currently are stored in IntegerFields, which is not enough when storing
file sizes of very large assets in bytes. A DataError integer out of range was raised.
Changing the according fields to BigIntegerFields solved this issue.