PB-1009: ignore invalid authentication header.
API Gateway always sets the `Geoadmin-Authenticated` and `Geoadmin-Username`
headers. When it is not able to authenticate the user, the values are "false"
and a single whitespace respectively.

The `PersistentRemoteUserMiddleware` class is designed to keep track of users
that were authenticated through headers but if the header is set with an
invalid value it will log out the user.

This change updates our derivation of `PersistentRemoteUserMiddleware` such that
it ignores the `Geoadmin-Username` header if `Geoadmin-Authenticated` is not set
to "true".

A ticket was opened with upstream to make `PersistentRemoteUserMiddleware`
more flexible in that regard: https://code.djangoproject.com/ticket/35971

This change also add tests exercising this new behaviour.