PB-1293 Rename "items" to "features"
This makes the naming consistent with the payload of getFeatures. The confusing part is that getFeatures responds to endpoint /collections/{collectionId}/items and not .../features.

Probably all should use "items" to be in line with the STAC terminology. "Features" is used because of the "FeatureCollection" within the geojsons that are delivered.

We choose "features" not because it's the best term but because it's consistent with getFeatures.