PB-1091 Flag the inexistent files in the file size updater
For some assets the file on the bucket is missing. This makes the
file_size_upgrade cronjob be stuck in a infinite loop, as it'll always
try to update the next N assets.
Therefore we flag these assets with a `file_size` of `None`, to avoid
repeating the update and to be able to produce a list of missing files.