Make spec building deterministic.
In order for yq to produce deterministic output, the order of its input files
needs to be deterministic. The order of the GNU Make prerequisites exposed with
the $^ automatic variable is not deterministic.

This updates the Makefile to sort the yq input files, hence making the rule
produce deterministic output. This also regenerates the spec based on the new rule.