Appease GitGuardian.
The [GitGuardian Bearer token detector](https://docs.gitguardian.com/secrets-detection/secrets-detection-engine/detectors/generics/bearer_token)
does not like it when we add a realistic-looking token. The documentation
shows a few ways to make it ignore that token. In this change we use the
dummy "123456" value.

The previous token was a more realistic dummy but a dummy nonetheless. It has
never been a valid token and could not be used to access anything.