PB-1279: Skip payload validation if no initial_data
There is a `Serializer.get_initial()` that would do something similar but not quite. If we use `get_initial()` , the test for the new POST endpoint works but other existing tests fail, like e.g. `ItemsUpdateEndpointTestCase.test_item_endpoint_patch_extra_payload`.

Not fully clear at this point why. Also, it's not clear if this renders the `validate_json_payload` useless for the new serializer (`ItemListSerializer`).

What is clear: `initial_data` is not set for AssetBaseSerializer when calling `ItemListSerializer.is_valid()`. So `validate_json_payload()` fails.

The serializers are connected like this:

   ItemLIstSerializer
   --> ItemSerializer
   --> AssetsForItemSerializer
   --> AssetBaseSerializer