PB-1279: Mark update() as not implemented
Contrary to the initial plan to make the POST /collections/{collection_id}/items endpoint a place to both create and update ("upsert") items, it is now reduced to only create new items. So we can address the "abstract-method" warning by the linter by just expressing that we don't intend to use that functionality.