PB-1279: Set itemAssets as writeable
Otherwise for bulkCreateItems the assets do not show up in the example payload.

The endpoints making use of itemAssets:

/collections
    GET getCollections
/collections/{collectionId}
    GET describeCollection
    PUT updateCollection
    PATCH partialUpdateCollection
/collections/{collectionId}/items/{featureId}
    GET getFeature
    PUT putFeature
    PATCH patchFeature
    PUT putFeature
/collections/{collectionId}/items
    GET getFeatures
    POST bulkCreateItems
/search
    GET getSearchSTAC
    POST postSearchSTAC

However, this is not visible in the specs because all writing endpoints have hardcoded examples instead of relying on their components. Also, we agreed that being able to write assets also elsewhere is not necessarily a problem.