PB-1331: Set default file path for asset
When assets are created via admin GUI there is no file path set. If the file is
then uploaded via API using the multipart upload, the file path will never be
set on the asset. This aligns the admin GUI behaviour with that of creating an
asset via API PUT. It has the downside that newly created assets already have
a file path that is a link to a non-existent file.