PB-1440: allow setting SESSION_* settings via environment variables.
We want to tweak those but we don't need to hardcode the exact values for all
environments immediately. Another change will update the values in DEV, INT
then PROD. Afterwards we can consider setting these values by default in STAC
itself.

This also removes the specific values we set for the dev build.