PB-1469: provide checksum for v1 and v0.9 API.
In v1 we provide the checksum through the `file:checksum` field. In v0.9 that field is called `checksum:multihash`. With this change we provide both fields.

We could provide just the field for v0.9 but then we need to remember to update it again when switching back to v1. Given how this code is not really tested automatically, we make the switch a little easier to get right by providing both fields.