PB-1495 Remove date comparison for property `expires`
In the forecast extension, `datetime` is the forecast date. So `datetime` can be anywhere in the future and `expires` might be before that. So comparing these two dates makes no sense. Same for `end_datetime`.

The `expires` field is part of the Timestamps extension, so it might be used in different contexts where this check would actually make sense. But ultimately, it is the responsibility of the data owner that the data is consistent. So we remove the check to reduce complexity on our side.