PB-124: Made the asset ID and content-type field read only in admin interface
The asset ID should be kept in sync with the asset data file name on S3, same
as for the asset content-type that should be kept in sync on S3 file.

Because we cannot change an S3 file name nor change an S3 file property (e.g.
content-type) without re-uploading it, we don't allow those changes on the
admin interface.