PB-1495 Fix tests that failed because `expires` is in the past
As we validate now that property `expires` is not in the past, we have to adapt all tests that create such objects.

In order to  still be able to create "expired" objects, we do the following:

1. Create an object with `expired` set to a point in time in the very near future (e.g. now() + 10ms).
2. Pause the thread for a time which is longer than the expiration limit.
3. Now there should be expired objects present.

I have my doubts that this is the best solution. It at least avoids changing the interface further just for testing.