PB-1495 Patch `timezone.now()` instead of using `sleep()`
Some tests check that expired objects are indeed removed. As property `expires` can no longer be in the past, these tests need to be adapted.

In a first try, objects were set to expired by putting the thread to sleep. That is rather shaky, so, instead, we now artificially advance the clock by patching `django.utils.timezone.now()` which is used in `validators_view.validate_item()`.

This also allows keep closer to the original tests in `test_remove_expired_items`.