PB-1519 Mark assets as external in bulk upload
This prevents having asset's href field to get an accidental prefix with the internal hosting address (on INT: https://sys-data.int.bgdi.ch/). So for example "https://sys-data.int.bgdi.ch/https://rgw.cscs.ch/...".

The serializer field `AssetBaseSerializer.href` checks for this flag and creates an internal URL (composed of the item/asset name) if the `Asset.is_external` is set.

For simplicity, we currently set all assets as external for the bulk upload endpoint. We might want to allow separate upload as well, as it's done for the `putAsset` endpoint.