PB-1519 Skip href reachability check for bulk upload
This is a temporary solution to unblock our users. Checking reachability of all assets in a bulk upload would introduce a significant performance penalty if done sequentially.

The`context` field is used to propagate the flag down to the `AssetBaseSerializer`. That does not look ideal, as the `context` field  is intended to pass along runtime information. But we know already at compile time if we want to check reachability.

Using a classic class field on a Serializer has proven to be challenging because we are using lists (so the serializer is called with `many=True`). So I resorted to the `context` field as a minimally invasive change. The proper way would probably be to create dedicated Serializers.

We might want to launch these requests to test reachability asychronously to still make this work in reasonable time. A short test with asyncio demonstrates that this might finish in max 0.6s for 100 assets. We decided to revisit this later.