PB-1557 Use aiohttp for href reachability
This is a first step towards enabling the bulk upload to do an asynchronous reachability check. The goal is to use the same reachability check for both the regular, synchronous checks for single assets as for the asychronously launched checks for multiple assets in the bulk upload.

aiohttp is chosen over httpx because it has a narrower focus. It's about enabling fast HTTP requests while httpx offers more convenience, similar to "requests". It would be possible to implement the HTTP request ourselves with asyncio but aiohttp seems to be a very established library exactly for that.

There is an initialization overhead for the single assets that comes with running a single reachability check in an event loop. The overhead is probably small (that is the idea of coroutines after all) but I have not tested its impact on runtime.

In order to be able to mock the HTTP request sent by aiohttp in the unit tests, aioresponses is introduced as a third party dependency. This replaces library "responses" which is no longer needed, as we replace "requests" with "aiohttp". Library "requests" is still in the Pipfile as it is used in a unit test elsewhere (`write_perftests.py`).