PB-1557 Use gevent instead of asyncio
Instead of launching another event loop with asyncio/aiohttp, use `gevent` + `requests` for async IO. Networking library `gevent` is already part of our stack through `gunicorn`.

That means reverting most of the changes of this branch. All uses of `asyncio` and `aiohttp`, and  `aioresponses` for testing, are removed again. We had already used `requests` for the original synchronous reachability check. We keep the async reachability check in `ItemListSerializer` and the URL in the error message.