PB-1531: only dump stacks when exit is taking too long.
In https://github.com/geoadmin/service-stac/pull/558 we added the ability to
log stack traces upon exit. However we did not distinguish between a normal
graceful exit and an abnormal exit of a deadlocked service. This means that on
every exit/restart, we clutter the logs with stack traces.

With this change, upon exit we only attempt to dump the stack if the worker is
still running after a delay. This means that graceful exit that complete within
that delay do not clutter the logs with stack traces.

This change removes the `GUNICORN_DUMP_STACKS_ON_EXIT` variable and adds
`GUNICORN_STACK_DUMP_DELAY` and `GUNICORN_GRACEFUL_TIMEOUT`. The former being
set by default to one second less than the latter. We may need to tweak
`GUNICORN_GRACEFUL_TIMEOUT` to be coherent with the Kubernetes
`terminationGracePeriodSeconds`.

Another way to resolve this might be implemented upstream later in https://github.com/benoitc/gunicorn/issues/3385