PB-1557 Add other validations to href field
To have the validation properly per assset, you would ahve to put this into `AssetsForItemSerializer.serialize()`. But this is also used for the PUT endpoint at `/collections/{collectionId}/items/{itemId}` through view `ItemDetail`.

It's not clear for the moment whether it's a bug or not that this other PUT endpoint does not validate the assets' href.