Create test Items and Assets in bulk.
The data_factory creates Django objects individually, which is really slow when
many objects are involved as that incurs one database round-trip per object.

With this change we use the `bulk_create` method instead. This dramatically
speeds up the test setup and allows much more practical performance testing
of larger datasets.

Also bump the number of Items in RemoveExpiredItemsManyWithProfiling to 1010.

Profiling output on my laptop:

	memory consumed: 6857881 (6857.881 per item)
	time consumed: 15.571250406s (0.015571250406000001 per item)