Add --max-deletions* safety flags.
This adds two flags: --max-deletions (absolute value) and
--max-deletions-percentage (a percentage). If we detect an attempt to remove
more than either value of Items, we fail with an error.

This is useful to guard against bugs that would accidentally wipe out the
whole database and to reduce risks of overloading the database with a request
that might consume a lot of resources.

Default values are based on the daily usage we observe in PROD. Given the
backlog of expired Items currently in the database, the safety will trigger
until we clean things up manually.

This also forbids negative values for --min-age-hours.