PB-1575: make deletion batch size configurable.
One `delete()` per object is very slow. One `delete()` for all the objects
consumes a lot of memory and fails in an all-or-nothing mode (i.e. when it
fails mid-way, nothing gets deleted).

With this change, we split the objects into batches of configurable size and
run `delete()` over each batch. It consumes a ~constant amount of memory, is
faster than having a batch size of 1 and is more robust. I.e. when it fails
mid-way, it will most likely have deleted at least some objects and the next
run has that much less work to do.

There will be cases where expired items are partially deleted (i.e. some of
their assets are deleted but not all). I don't expect that to be a problem for
our use case.