PB-1788: add way to slow down checker endpoint.
We want to add a readiness probe and update the liveness probe to be more
tolerant of slow responses. In order to verify this works as expected, we need
to be able to inject arbitrary delays in checker responses.

This change adds a `CHECKER_DELAY` setting that allows delaying the `/checker`
endpoint by a commensurate number of seconds. This also adds previously missing
unit test against the checker endpoint.