PB-1911 Add publiccode.yml
This was created with a Beta version of the [publiccode.yml Editor](https://puzzle.github.io/publiccode-editor/) of the Federal Chancellery. The idea is that with such a file, the service-stac repo is going to be picked up by a crawler populating the [OSS Catalogue](https://opensource.admin.ch) of the Swiss Confederation.

The description was adapted from the Tech Docs [Overview page of the STAC API](https://sys-docs.int.bgdi.ch/docs/stac-api/overview.html) (still in Beta).

I was not sure what to list as featues, these are just from the top of my head.

The `softwareVersion` and `releaseDate` ([optional fields](https://yml.publiccode.tools/schema.core.html#key-softwareversion)) were left out on purpose because it would require us to adapt the CI to keep this up-to-date.